+++ R/SimDesign.R
@@ -34,6 +34,7 @@
 #' @importFrom future.apply future_lapply
 #' @importFrom progressr progressor
 #' @importFrom beepr beep
+#' @importFrom snow sendCall
 # @importFrom robustbase glmrob
 #' @importFrom utils recover packageVersion head tail capture.output object.size
 #' @keywords package