+++ R/SimCheck.R
@@ -4,9 +4,11 @@
 #' evaluation check whether all \code{.rds} files have been saved. If missing
 #' the missing row condition numbers will be returned
 #'
+#' @param dir character vector input indicating the directory
+#'   containing the \code{.rds} files (see \code{files})
+#'
 #' @param files vector of file names referring to the saved simulation files.
-#'   E.g. \code{c('mysim-1.rds', 'mysim-2.rds', ...)}. Default assumes
-#'   that all the files are available in the current working directory
+#'   E.g. \code{c('mysim-1.rds', 'mysim-2.rds', ...)}
 #'
 #' @param min minimum number after the \code{'-'} deliminator. Default is 1
 #'
@@ -28,13 +30,24 @@
 #' @examples
 #' \dontrun{
 #'
+#' # if files are in mysimfiles/ directory
+#' SimCheck('mysimfiles')
+#'
+#' # specifying files explicility
 #' setwd('mysimfiles/')
-#' files <- dir()
-#' SimCheck()
+#' SimCheck(files=dir())
 #'
 #' }
 #'
-SimCheck <- function(files = dir(), min = 1L, max = NULL){
+SimCheck <- function(dir = NULL, files = NULL, min = 1L, max = NULL){
+    if(is.null(dir) && is.null(files))
+        stop('either dir or files must be specified')
+    if(!is.null(dir) && !is.null(files))
+        stop('dir OR files must be specified, not both')
+    if(!is.null(dir)){
+        files <- dir(dir)
+    } else dir <- './'
+    files <- paste0(dir, files)
     filename <- strsplit(files[1], '-')[[1L]][1L]
     if(is.null(max)){
         subfiles <- gsub(paste0(filename, '-'), files, replacement = '')